backend/requirements.txt
Flask==3.0.3
pymysql==1.1.1
bcrypt==4.2.0
python-dotenv==1.0.1
scikit-learn
joblib
requests

